completed backend API

+++ mern-app/backend/controllers/goalController.js
@@ -1,9 +1,10 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 // GET 
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
     res.status(200).json(goals)
 })
 // POST
@@ -13,7 +14,8 @@ const setGoal = asyncHandler(async (req, res) => {
         throw new Error('Please add a text field')
     }
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id,
     }) 
     res.status(200).json(goal)
 })
@@ -24,7 +26,21 @@ const updateGoal = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('Goal not found')
     }
-    const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, { new: true })
+
+    const user = await User.findById(req.user.id)
+    //Check for user
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    //Make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
+    const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, { new: true, })
     
     res.status(200).json(updatedGoal)
 })
@@ -35,6 +51,18 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if (!goal) {
         res.status(400)
         throw new Error(`Goal not found`)
+    }
+    const user = await User.findById(req.user.id)
+    //Check for user
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    //Make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
     await Goal.findByIdAndDelete(req.params.id)
     res.status(200).json({id: req.params.id})
